+++ internal/provider/security/security_account_resource_test.go
@@ -16,10 +16,19 @@ func TestAccSecurityAccountResource(t *testing.T) {
 		Steps: []resource.TestStep{
 			// Test updating a resource
 			{
-				Config: testAccSecurityAccountResourceConfig("tf_acc_test", "password123"),
+				Config: testAccSecurityAccountResourceConfigCreate("tf_acc_test", "password123"),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("netapp-ontap_security_account.security_account", "name", "tf_acc_test"),
 					resource.TestCheckResourceAttr("netapp-ontap_security_account.security_account", "password", "password123"),
+				),
+			},
+			// Test updating a resource with comment and locked
+			{
+				Config: testAccSecurityAccountResourceConfig("tf_acc_test", "update", true),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("netapp-ontap_security_account.security_account", "name", "tf_acc_test"),
+					resource.TestCheckResourceAttr("netapp-ontap_security_account.security_account", "comment", "update"),
+					resource.TestCheckResourceAttr("netapp-ontap_security_account.security_account", "locked", "true"),
 				),
 			},
 			// Test importing a resource
@@ -35,7 +44,43 @@ func TestAccSecurityAccountResource(t *testing.T) {
 	})
 }
 
-func testAccSecurityAccountResourceConfig(name string, accpassword string) string {
+func testAccSecurityAccountResourceConfig(name string, comment string, locked bool) string {
+	host := os.Getenv("TF_ACC_NETAPP_HOST")
+	admin := os.Getenv("TF_ACC_NETAPP_USER")
+	password := os.Getenv("TF_ACC_NETAPP_PASS")
+	if host == "" || admin == "" || password == "" {
+		fmt.Println("TF_ACC_NETAPP_HOST, TF_ACC_NETAPP_USER, and TF_ACC_NETAPP_PASS must be set for acceptance tests")
+		os.Exit(1)
+	}
+	return fmt.Sprintf(`
+provider "netapp-ontap" {
+ connection_profiles = [
+    {
+      name = "cluster2"
+      hostname = "%s"
+      username = "%s"
+      password = "%s"
+      validate_certs = false
+    },
+  ]
+}
+
+resource "netapp-ontap_security_account" "security_account" {
+  # required to know which system to interface with
+  cx_profile_name = "cluster2"
+  name = "%s"
+  applications = [{
+    application = "http"
+    authentication_methods = ["password"]
+  }]
+  comment = "%s"
+  locked = %t
+  password = "password123"
+}
+`, host, admin, password, name, comment, locked)
+}
+
+func testAccSecurityAccountResourceConfigCreate(name string, accpassword string) string {
 	host := os.Getenv("TF_ACC_NETAPP_HOST")
 	admin := os.Getenv("TF_ACC_NETAPP_USER")
 	password := os.Getenv("TF_ACC_NETAPP_PASS")